Move embedding loading to utils
