+++ examples/single_cell_embedding/src/scembed/methods/base.py
@@ -272,31 +272,6 @@ class BaseIntegrationMethod(ABC):
         logger.info("Saved %s embedding to '%s'", self.name, file_path)
         return file_path
 
-    @staticmethod
-    def load_embedding(file_path: Path | str) -> pd.DataFrame:
-        """Load a saved embedding file back into a DataFrame."""
-        file_path = Path(file_path)
-        if not file_path.exists():
-            raise ValueError(f"File does not exist: {file_path}")
-
-        if file_path.suffix == ".parquet":
-            return pd.read_parquet(file_path)
-
-        elif file_path.name.endswith(".pkl.gz"):
-            with gzip.open(file_path, "rb") as f:
-                return pickle.load(f)
-
-        elif file_path.suffix == ".h5":
-            with h5py.File(file_path, "r") as hf:
-                return pd.DataFrame(
-                    data=hf["embedding"][:],
-                    index=[n.decode() for n in hf["cell_names"][:]],
-                    columns=[n.decode() for n in hf["dim_names"][:]],
-                )
-
-        else:
-            raise ValueError(f"Unsupported file format: {file_path.name}. Supported: .parquet, .pkl.gz, .h5")
-
     def get_model_info(self) -> dict[str, Any]:
         """
         Get information about the fitted model.
